+++ GRDB/ValueObservation/DatabaseCancellable.swift
@@ -17,10 +17,14 @@ public class AnyDatabaseCancellable: DatabaseCancellable {
         _cancel = cancel
     }
     
-    /// Initializes a cancellable object that forwards cancellation to the
+    /// Creates a cancellable object that forwards cancellation to the
     /// provided cancellable.
     public convenience init(_ cancellable: DatabaseCancellable) {
-        self.init(cancel: cancellable.cancel)
+        var cancellable: DatabaseCancellable? = cancellable
+        self.init {
+            cancellable?.cancel()
+            cancellable = nil // Release memory
+        }
     }
     
     deinit {